Cache

add cache for api

+++ test.py
@@ -29,12 +29,18 @@ def cleanup():
     from models.site import Site, Block
     Site.drop_table(fail_silently=True)
     Block.drop_table(fail_silently=True)
-    from query.comment import local
-    for k, v in local.iteritems():
+    from utils.cache import local_cache
+    for k, v in local_cache.iteritems():
+        if not k.startswith('comment:'):
+            continue
         v.drop_table(fail_silently=True)
+    from utils.cache import rds
+    keys = rds.keys('comment:*')
+    if keys:
+        rds.delete(*keys)
 
 def generate_path():
-    block = ['init', 'test']
+    block = ['init', 'test', 'libs']
     path = os.path.realpath(os.path.dirname(__file__))
     for path in os.listdir(path):
         if not path.startswith('.') \
